LayoutTests/svg/dynamic-updates/SVGEllipseElement-svgdom-rx-prop-expected.txt

SVG 1.1 dynamic update tests

Tests dynamic updates of the 'rx' property of the SVGEllipseElement object

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS ellipseElement.rx.baseVal.value is 0
PASS ellipseElement.rx.baseVal.value is 100
PASS successfullyParsed is true

TEST COMPLETE
